Created a single base which all our view models extend
