+++ app/src/main/java/com/task4/task4/viewmodels/TaskListViewModel.kt
@@ -1,4 +1,4 @@
-package com.task4.task4
+package com.task4.task4.viewmodels
 
 import androidx.lifecycle.Transformations
 import androidx.lifecycle.ViewModel
@@ -6,22 +6,13 @@ import com.task4.task4.database.Task
 import com.task4.task4.database.TaskRepository
 import com.task4.task4.database.TaskWithSubTasks
 
-class TaskListViewModel : ViewModel() {
+class TaskListViewModel : BaseViewModel() {
 
-    private val taskRepository = TaskRepository.get()
     val taskListLiveData = Transformations.map(taskRepository.getTasksWithSubtasks()) {
         it.map { taskRepository.getTaskWithSubtasks(it.parent.id) }
     }
 
     fun addTask(task: Task) {
         taskRepository.createTask(task)
-    }
-
-    fun saveTasks(tasks: List<TaskWithSubTasks>) {
-        tasks.forEach { saveTask(it) }
-    }
-
-    fun saveTask(task: TaskWithSubTasks) {
-        taskRepository.update(task)
     }
 }
